Added option to set individual LED strahler only. Did some more prework for music and LJ control in the windows app.

+++ LED_Strahler_Win_App/LED_Strahler/LED_Strahler/LEDStrahler.cs
@@ -30,8 +30,8 @@
             set { _UUID = value; NotifyPropertyChanged(); }
         }
 
-        private int _Group = 0;
-        public int Group
+        private byte _Group = 0;
+        public byte Group
         {
             get { return _Group; }
             set
